add option to print config
New option to just print the configuration for feedback to the user.
If the config is not valid, the runner does what it did before and
exits with 2 after displaying the errors.

When the option is set,the linter prints the options, rules, and
paths it would use and exits with a 0 status.