Load schema asynchronously
This commit makes schema loading asynchonous so that a schema can be
read from stdin when the input source might block.

This commits separates schema loading from the config. The schema is now
passed to the config as an argument. Loading of the linter config file
has also been pulled out of the Configuration class because that config
can incluence how the schema is loaded.

The Configuration getSchema and getSchemaSourceMap functions have been
kept intact since rules may use these functions.

I've verified that the `test.js` file in #211 successfully runs with
this change.

Fixes #211.