Add support for autofixing
The core API is fairly simple: errors can have a new field, `fix`, which
includes a location (likely a graphql node's `.loc`) and a replacement.
The runner then takes care of applying the fix, if the `--fix` option
was specified.

I added autofixers for all of the capitalization and alphabetization
rules.  (The remainder can't (or shouldn't) really be autofixed.)  In
general, capitalization was very easy, and alphabetization was quite
tricky, but in the latter case it's mostly in a shared util that handles
all of annoying syntax-transformation junk.

Fixes #23.