Wait for stdout writing to finish before exiting
This fixes a race condition when writing to stdout on certain platforms and streams.

As noted in [A note on process I/O](https://nodejs.org/api/process.html#process_a_note_on_process_i_o) in the Node documentation, the `process.stdout` stream is asynchronous in some cases (pipes on POSIX, TTYs on Windows).

In these cases, the output sent to stdout can sometimes be cut off at the first chunk because the linter process exits before the stream is finished writing. The result is either a truncated set of errors (with default output) or invalid JSON (with JSON output) when the set of errors exceeds the default chunk size.

This is fixed by using the return status of `socket.write` to determine if some of the data was queued, then awaiting the socket's `'drain'` event if so before returning.