Check for `ruleName` instead of using `instanceof`
Custom rules written outside of `graphql-schema-linter` cannot create instances of the `ValidationError` class that's used inside `validator.js` because they have to import from their own reference to it, which is typically separate from the reference that the globally-installed `graphql-schema-linter` is using.

Instead of using an `instanceof` check (which requires the object reference to be identical), we check for the presence of `ruleName` for (mostly) equivalent behavior.

Tests are added that demonstrate and document the desired behavior.

Fixes #265.